feat(twitter): implement KnowledgeProcessor for tweet analysis
- Introduced the KnowledgeProcessor class to encapsulate the logic for processing knowledge users' tweets, enhancing modularity and separation of concerns.
- The TwitterInteractionClient now utilizes the KnowledgeProcessor to handle tweet analysis, improving code readability and maintainability.
- This change aligns with the goal of enhancing the functionality of the Twitter interaction logic while adhering to SOLID principles.
